Moves 'columns' specific init interfaces into base_tests.py, tidies up missing inherit tests, adds a BaseTransformer.transform() call to mapping transformer so that the relevant input checks are called
