+++ tests/base_tests.py
@@ -11,8 +11,14 @@ import tests.test_data as d
 
 class GenericInitTests:
     """
-    Generic tests for transformer.init().
-    Note this deliberately avoids starting with "Tests" so that the tests are not run on import.
+    Generic tests for transformer.init(). This test class does not contain tests for the behaviours
+    associated with the "columns" argument because the structure of this argument varies between
+    transformers. In this file are other test classes that inherit from this one which are specific
+    to the different "columns" argument structures. Please choose the appropriate one of them to inherit
+    when writing tests unless the transformer is a special case and needs unique tests written for
+    "columns", in which case inherit this class.
+
+    Note this class name deliberately avoids starting with "Tests" so that the tests are not run on import.
     """
 
     def test_print(self, instantiated_transformers):
@@ -63,6 +69,93 @@ class GenericInitTests:
                 copy=non_bool,
                 **minimal_attribute_dict[self.transformer_name],
             )
+
+
+class ColumnStrListInitTests(GenericInitTests):
+    """
+    Tests for BaseTransformer.init() behaviour specific to when a transformer takes columns as string or list.
+    Note this deliberately avoids starting with "Tests" so that the tests are not run on import.
+    """
+
+    def test_columns_empty_list_error(
+        self,
+        minimal_attribute_dict,
+        uninstantiated_transformers,
+    ):
+        """Test an error is raised if columns is specified as an empty list."""
+
+        args = minimal_attribute_dict[self.transformer_name].copy()
+        args["columns"] = []
+
+        with pytest.raises(ValueError):
+            uninstantiated_transformers[self.transformer_name](**args)
+
+    @pytest.mark.parametrize("non_string", [1, True, {"a": 1}, [1, 2], None])
+    def test_columns_list_element_error(
+        self,
+        non_string,
+        minimal_attribute_dict,
+        uninstantiated_transformers,
+    ):
+        """Test an error is raised if columns list contains non-string elements."""
+
+        args = minimal_attribute_dict[self.transformer_name].copy()
+        args["columns"] = [non_string, non_string]
+
+        with pytest.raises(
+            TypeError,
+            match=re.escape(
+                f"{self.transformer_name}: each element of columns should be a single (string) column name",
+            ),
+        ):
+            uninstantiated_transformers[self.transformer_name](**args)
+
+    @pytest.mark.parametrize("non_string_or_list", [1, True, {"a": 1}, None])
+    def test_columns_non_string_or_list_error(
+        self,
+        non_string_or_list,
+        minimal_attribute_dict,
+        uninstantiated_transformers,
+    ):
+        """Test an error is raised if columns is not passed as a string or list."""
+
+        args = minimal_attribute_dict[self.transformer_name].copy()
+        args["columns"] = non_string_or_list
+
+        with pytest.raises(
+            TypeError,
+            match=re.escape(
+                f"{self.transformer_name}: columns must be a string or list with the columns to be pre-processed (if specified)",
+            ),
+        ):
+            uninstantiated_transformers[self.transformer_name](**args)
+
+
+class ColumnsFromDictInitTests(GenericInitTests):
+    """
+    Tests for BaseTransformer.init() behaviour specific to when a transformer reads columns from a dict.
+    Note this deliberately avoids starting with "Tests" so that the tests are not run on import.
+    """
+
+    @pytest.mark.parametrize("non_string", [1, True, None])
+    def test_columns_list_element_error(
+        self,
+        non_string,
+        minimal_attribute_dict,
+        uninstantiated_transformers,
+    ):
+        """Test an error is raised if columns list contains non-string elements."""
+
+        args = minimal_attribute_dict[self.transformer_name].copy()
+        args["mappings"][non_string] = {1: 2, 3: 4}
+
+        with pytest.raises(
+            TypeError,
+            match=re.escape(
+                f"{self.transformer_name}: each element of columns should be a single (string) column name",
+            ),
+        ):
+            uninstantiated_transformers[self.transformer_name](**args)
 
 
 class GenericFitTests:
